feat(test_margin): split testcase to 3 files
